+++ NEWS.md
@@ -4,6 +4,7 @@
 
 # Version 2.7.1
 - Revision of recruitment model, with the addition of a recruitment probability (ProbRecr) within the bioclimatic limits
+- Temperature effects on sapwood conversion to heartwood
 
 # Version 2.7.0
 - Basic water balance model with relative water content